requirements.txt
openai==2.5.0
streamlit==1.35.0
requests==2.31.0
python-dotenv==1.0.1
sqlite-utils==3.36
pandas==2.2.2
numpy==1.26.4
pdfplumber==0.11.4
docx2txt==0.8
semantic-kernel==1.3.0
pydantic<2.0
azure-cli==2.56.0
python-docx==1.1.2
rapidfuzz>=3.5.0